+++ src/test/java/carracing/CarRacingGameTest.java
@@ -1,27 +1,21 @@
-package step3;
+package carracing;
 
+import carracing.controller.CarRacingGame;
+import carracing.domain.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class CarRacingGameTest {
-
-    @Test
-    void play() {
-        CarRacingGame game = new CarRacingGame(new String[]{"pobi", "crong", "honux"});
-        assertThatThrownBy(() -> game.play(0)).isInstanceOf(RuntimeException.class)
-                .hasMessageContaining("최소 한 번이상 시도해야합니다.");
-    }
 
     @Test
     @DisplayName("우승자들의 현위치가 최대거리가 맞는지 확인")
     void getWinners() {
         CarRacingGame game = new CarRacingGame(new String[]{"pobi", "crong", "honux"});
-        game.play(4);
+        game.play();
 
         List<Car> winners = game.getWinners();
         int maxDistance = winners.get(0).getLocation();
