+++ src/test/java/carracing/CarTest.java
@@ -1,0 +1,14 @@
+package carracing;
+
+import carracing.domain.Car;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class CarTest {
+
+    @Test
+    void car_init_location() {
+        assertThat(new Car("name").getLocation()).isZero();
+    }
+}
